Refactor/instructions page

+++ src/scripts.js
@@ -55,10 +55,6 @@ window.addEventListener("load", () => {
     });
 });
 
-instructionsButton.addEventListener("click", () => {
-  hide(instructionsPage);
-});
-
 loginForm.addEventListener("submit", e => {
   e.preventDefault();
 
@@ -140,13 +136,14 @@ userSearchForm.addEventListener("submit", e => {
   show(userSpentHeader);
 });
 
-upcomingMinis.addEventListener('click', e => {
-  if (e.target.id !== 'upcomingMinis') {
+upcomingMinis.addEventListener("click", e => {
+  if (e.target.id !== "upcomingMinis" && e.target.id !== "deleteReservation") {
     let currentMiniCard = e.target.parentNode;
-    let deleteButton = currentMiniCard.children[1]
+    let deleteButton = currentMiniCard.children[2]
     hide(currentMiniCard.children[0]);
-    show(currentMiniCard.children[1]);
-    deleteButton.addEventListener('click', () => {
+    hide(currentMiniCard.children[1]);
+    show(currentMiniCard.children[2]);
+    deleteButton.addEventListener("click", () => {
       deleteBooking(currentMiniCard.id);
     });
   }
@@ -272,8 +269,8 @@ function generateReservations(bookings) {
     upcomingMinis.innerHTML += `
     <div class="mini-room-booked" id="${reservation.id}">
       <h3 tabindex="0">Room ${reservation.roomNumber}</h3>
-      <button class="delete-reservation hidden"id="deleteReservation">Delete.</button>
       <h3 tabindex="0">${reservation.date}</h3>
+      <button class="delete-reservation hidden"id="deleteReservation">Delete.</button>
     </div>
     `;
   });
@@ -340,9 +337,9 @@ function logout() {
 
 function unsuccessfulLogin(reason) {
   if (reason === "password") {
-    loginErrorText.innerText = "Please enter a valid password";
+    loginErrorText.innerText = `Password should be "overlook2021".`;
   } else if (reason === "username") {
-    loginErrorText.innerText = "Please enter a valid username";
+    loginErrorText.innerText = `Please enter a valid username starting with "customer" and ending with 1-50. eg "customer1", "customer7", "customer25"`;
   }
   show(loginErrorText);
   loginErrorText.classList.add("shake");
@@ -350,7 +347,7 @@ function unsuccessfulLogin(reason) {
     hide(loginErrorText);
     resetLoginForm();
     loginErrorText.innerText = "";
-  }, 1000);
+  }, 5000);
 }
 
 function convertDateDashes(date) {
